progress spinner moved to common components
