+++ app/renderer/components/common/progress_spinner.vue
